requirements.txt
requests
python-dotenv
playwright
python-telegram-bot==13.15
fastapi
uvicorn[standard]
